Mise a jour depot svn

+++ trunk/Annulation.php
@@ -5,41 +5,42 @@
   {
 		header("Location: LoginClub.php");
   }
-
   require "partage.php";
-  xhtml_pre("Annullation d'une inscription");
-?>
 
-
-<div >
-
-<?php
-try
-{
+  try
+  {
 	// On se connecte à MySQL
     $pdo_options[PDO::ATTR_ERRMODE] = PDO::ERRMODE_EXCEPTION;
 	//$bdd = new PDO('mysql:host=localhost;dbname=LASER', 'root', 'root', $pdo_options);
     $bdd = new PDO($pdo_path, $user, $pwd, $pdo_options);
     
-    $req = $bdd->prepare('SELECT nom , prenom FROM Inscrit WHERE ID_inscrit = ?');
+    $req = $bdd->prepare('SELECT nom, prenom, ID_regate FROM Inscrit WHERE ID_inscrit = ?');
 	$req->execute(array($_GET['ID']));
     $donnees = $req->fetch();
     
-    if(isset($donnees['prenom']))
+    if(
+      isset($donnees['prenom']) // si nous avons trouvé le coureur
+      and
+      $donnees['ID_regate'] == $_SESSION['ID_regate']   // nous en avons les droits
+    ) 
     {
 
     	$sql = 'DELETE FROM Inscrit WHERE ID_inscrit =?';
    		$req->closeCursor();
 		$req = $bdd->prepare($sql);
 		$req->execute(array($_GET['ID']));
-	    
-	    echo $donnees['prenom'].' '.$donnees['nom'];
-		echo ' : inscription annulé';
-
     }
     else
     {
-    	echo 'ERREUR';
+        xhtml_pre("Annullation d'une inscription");
+    	echo '<div>ERREUR : coureur pas trouvé dans cette régate</div>';
+    
+        echo '
+        <p></p> 
+        <div>
+        Retour à la page de <a href="Regate.php">gestion de la régate</a>.
+        </div>';
+        xhtml_post();
     }
     
 
@@ -54,16 +55,7 @@ catch(Exception $e)
     die('Erreur : '.$e->getMessage());
 }
 
-?>
-
-</div>
-
-<p></p>
-<div>
-Retour à la page de <a href="Regate.php">gestion de la régate</a>.
-</div>
+header("Location: Regate.php?item=inscrits");
 
 
-<?php
-xhtml_post();
 ?>
